Add a dummy entry when a odd-numbered-length list of IDs is passed to generate_pairs
When an odd number of IDs was passed to generate_pairs, it would cause an error (OneFactorization::InvalidLengthError). So, we now ensure the list that is one_factorized is always even, by adding a dummy 'Flying solo' list member, so that the odd person out is matched with 'flying solo' and nothing breaks.